# static/swagger.yml
openapi: 3.0.0
paths:
...
  /api/v2/editor:
    post:
      operationId: editors-get
      summary: Получение списка редакторов
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorsGetRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorsGetResponse"
        "400": &ref_21
          description: Ответ будет таким если с данным что-то не так
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401": &ref_22
          description: Ответ для незарегистрированного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403": &ref_23
          description: Ответ для неавторизованного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404": &ref_24
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "406": &ref_25
          description: Не принято значение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
        "409": &ref_26
          description: Ответ для конфликта файлов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "412": &ref_27
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionFailedError"
        "500": &ref_28
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        "501": &ref_29
          description: Пока не реализовано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
      tags: &ref_30
        - editor
      security: &ref_31
        - bearer: []
    put:
      operationId: editor-create
      summary: Создание редактора
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorCreateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorGetResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
  /api/v2/editor/{editorId}:
    get:
      operationId: editor-get
      summary: Получение редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorGetResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
    patch:
      operationId: editor-update
      summary: Изменить редактор
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorUpdateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorGetResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
    delete:
      operationId: editor-delete
      summary: Удаление редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
  /api/v2/editor/layer/{editorId}:
    put:
      operationId: editor-layer-create
      summary: Создание слоя редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorLayerCreateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorLayerGetResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
  /api/v2/editor/layer/{editorId}/{layerId}:
    get:
      operationId: editor-layer-get
      summary: Получение слоя редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
        - name: layerId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorLayerGetResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
    patch:
      operationId: editor-layer-update
      summary: Изменить слой редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
        - name: layerId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorLayerUpdateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorLayerGetResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
    delete:
      operationId: editor-layer-delete
      summary: Удаление слоя редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
        - name: layerId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
  /api/v2/editor/layer/{editorId}/{layerId}/{moveIndex}:
    put:
      operationId: editor-layer-move
      summary: Изменить очередь слоя редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
        - name: layerId
          required: true
          in: path
          schema:
            type: string
        - name: moveIndex
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
  /api/v2/editor/frame/{editorId}/{time}:
    post:
      operationId: editor-frame-get
      summary: Получение кадра из редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
        - name: time
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: Успешный ответ
          content:
            image/jpeg:
              encoding:
                image_jpeg:
                  contentType: image/jpeg
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
  /api/v2/editor/export/{editorId}:
    get:
      operationId: editor-export-status
      summary: Узнать статус экспорта видео из редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorGetRenderingStatusResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
    post:
      operationId: editor-export
      summary: Экспорт видео из редактора
      parameters:
        - name: editorId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorExportRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorGetRenderingStatusResponse"
        "400": *ref_21
        "401": *ref_22
        "403": *ref_23
        "404": *ref_24
        "406": *ref_25
        "409": *ref_26
        "412": *ref_27
        "500": *ref_28
        "501": *ref_29
      tags: *ref_30
      security: *ref_31
...
info:
  title: MyScreen backend API
  description: MyScreen backend API
  version: 2.41.29
  contact:
    name: Stanislav V Vyaliy
    url: https://wisekaa.dev
    email: stanislav@wisekaa.dev
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Токен авторизации
      name: token
  schemas:
    BadRequestError:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        code:
          type: string
          example: server-error.10004
        message:
          type: string
          example: Bad request
      required:
        - statusCode
        - code
        - message
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 401
        code:
          type: string
          example: server-error.10001
        message:
          type: string
          example: Unauthorized request
      required:
        - statusCode
        - code
        - message
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: number
          example: 403
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: Forbidden
      required:
        - statusCode
        - code
        - message
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        code:
          type: string
          example: server-error.10005
        message:
          type: string
          example: Not Found
      required:
        - statusCode
        - code
        - message
    NotAcceptableError:
      type: object
      properties:
        statusCode:
          type: number
          example: 406
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Not Acceptable
      required:
        - statusCode
        - code
        - message
...
    ConflictError:
      type: object
      properties:
        statusCode:
          type: number
          example: 409
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Conflict exists
        data:
          $ref: "#/components/schemas/ConflictData"
      required:
        - statusCode
        - code
        - message
        - data
    PreconditionFailedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 412
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: User exists
      required:
        - statusCode
        - code
        - message
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: number
          example: 500
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    NotImplementedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 501
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    Status:
      type: string
      enum:
        - success
        - error
...
    SuccessResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
      required:
        - status
...
    EditorsGetRequest:
      type: object
      properties:
        where:
          description: Запрос
          allOf:
            - $ref: "#/components/schemas/EditorPartialRequest"
        select:
          type: array
          description: Выбрать поля
          example: []
          items:
            type: string
            enum:
              - id
              - name
              - width
              - height
              - fps
              - renderingStatus
              - renderingPercent
              - renderingError
              - renderedFile
              - keepSourceAudio
              - totalDuration
              - videoLayers
              - audioLayers
              - createdAt
              - updatedAt
        scope:
          description: Рамки для запроса
          allOf:
            - $ref: "#/components/schemas/LimitRequest"
...
    EditorsGetResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        count:
          type: number
          description: Количество редакторов
        data:
          description: Оплаты
          type: array
          items:
            $ref: "#/components/schemas/EditorResponse"
      required:
        - status
        - count
        - data
    EditorCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Имя редактора
          example: имя редактора
        width:
          type: integer
          description: Ширина редактора
          example: 1920
        height:
          type: integer
          description: Высота редактора
          example: 1080
        fps:
          type: integer
          description: Фреймрейт
          example: 24
        keepSourceAudio:
          type: boolean
          description: Воспроизводить музыку с видео
          example: true
      required:
        - name
        - width
        - height
        - fps
        - keepSourceAudio
    EditorGetResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        data:
          description: Редактор
          allOf:
            - $ref: "#/components/schemas/EditorResponse"
      required:
        - status
        - data
    EditorUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Имя редактора
          example: имя редактора
        width:
          type: integer
          description: Ширина редактора
          example: 1920
        height:
          type: integer
          description: Высота редактора
          example: 1080
        fps:
          type: integer
          description: Фреймрейт
          example: 24
        keepSourceAudio:
          type: boolean
          description: Воспроизводить музыку с видео
          example: true
    EditorLayerCreateRequest:
      type: object
      properties:
        index:
          type: integer
          description: Индекс файла
          example: 1
          default: 1
        duration:
          type: number
          description: Длительность
          example: 10
          default: 10
        cutFrom:
          type: number
          description: С какой секунды начать вырезать клип
          example: 0
          default: 0
        cutTo:
          type: number
          description: До какой секунды вырезать клип
          example: 10
          default: 10
        start:
          type: number
          description: С какой секунды начинать воспроизводить клип
          default: 0
          example: 0
        mixVolume:
          type: integer
          description: Аудио дорожка из видео, 0-выключен, 1-включен
          example: 1
          default: 1
        file:
          type: string
          description: Файл
          format: uuid
      required:
        - file
    EditorLayerGetResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        data:
          description: Слой редактора
          allOf:
            - $ref: "#/components/schemas/EditorLayerResponse"
      required:
        - status
        - data
    EditorLayerUpdateRequest:
      type: object
      properties:
        index:
          type: integer
          description: Индекс файла
          example: 1
          default: 1
        duration:
          type: number
          description: Длительность
          example: 10
          default: 10
        cutFrom:
          type: number
          description: С какой секунды начать вырезать клип
          example: 0
          default: 0
        cutTo:
          type: number
          description: До какой секунды вырезать клип
          example: 10
          default: 10
        start:
          type: number
          description: С какой секунды начинать воспроизводить клип
          default: 0
          example: 0
        mixVolume:
          type: integer
          description: Аудио дорожка из видео, 0-выключен, 1-включен
          example: 1
          default: 1
...